Improved folder architecture for register agent
