+++ public/controllers/agent/register-agent/index.ts
@@ -1,0 +1,2 @@
+export { RegisterAgent } from './main'
+export * from './services'
